LibWeb: Reduce CSS parser token memory usage
Store CSS token payloads in a variant so each token only carries the
state needed by its type. Keep delimiter, number, hash, string, and
dimension data separate instead of storing every possible payload on
every token.

Use a smaller component-value token for function and block boundary
metadata. These component values only need token type, original source
text, and source positions, so avoid embedding full token payload
storage inside every Function and SimpleBlock.

Shrink CSS source positions to explicit 32-bit counters. Guard the C++
and Rust tokenizer paths against overflow. Add size assertions for the
hot Token and ComponentValue types so future growth is intentional.

+++ Libraries/LibWeb/CSS/Parser/SourcePosition.h
@@ -9,6 +9,6 @@
 #include <AK/Types.h>
 
 struct SourcePosition {
-    size_t line { 0 };
-    size_t column { 0 };
+    u32 line { 0 };
+    u32 column { 0 };
 };
